♻️ Move logger.go to public and add aliases
