+++ public_logger.go
@@ -1,0 +1,14 @@
+package sqlslog
+
+import "github.com/akm/sql-slog/public"
+
+type (
+	SqlLogger = public.SqlLogger
+)
+
+var (
+	NewSqlLogger = public.NewSqlLogger
+
+	IgnoreAttr  = public.IgnoreAttr
+	WithNilAttr = public.WithNilAttr
+)
